Change units to float
Moved classes to appropriate directories

+++ game/core/src/ntnu/gruppe22/game/helpers/GameInfo.java
@@ -6,8 +6,8 @@ package ntnu.gruppe22.game.helpers;
 
 public class GameInfo {
 
-    public static final int WIDTH = 630;
-    public static final int HEIGHT = 480;
-    public static final float PPM = 100;
+    public static final float WIDTH = 630.0f;
+    public static final float HEIGHT = 480.0f;
+    public static final float PPM = 100.0f;
 
 }
